fix: missing networks in non-emulated compose, assets manager permissions (#12)
* The web_network was missing from the collector in non emulated setups
* When the volume is created by Docker daemon on first run, it's created as root, so the container has a permission denied.

There's 2 ways to solve this:
    * Create the folder before spinning up the compose
    * Set the container's to root, albeit the increase in the surface of potential attacks.

In a discussion with @alvarobernal2412, we decided it's better to take option 2, since it's more maintainable to keep all the volume handling with Docker and not mess with it manually because we're probably going to change where the mountpoints of the container are located at some point (and that would require editing the script as well). The increase of the surface of attack is negligible in our use case.